Error List Implementation
Includes functions and devtools functionality for interacting with the error queue. Note that heap size was changed from 0, which was causing the malloc bugs from before. BEFORE RUNNING ANY ERROR RELATED CODE: make sure that your heap size is not set to 0 in Project > Properties > Arm Linker > Basic Options or it WILL NOT work.