MDE cycle; Error Seeding; SRAM Fixes
Added the MDE cycle to the debug menu and tested it; with integration and testing with the timer it should be good to go. Timer will have to be longer than a cycle, which is about 8 minutes thanks to the Flash.

Fixed error seeding for all chips; it now works properly.

SRAM had a bug where it was outputting a 0, this has been fixed. The program was previously incapable of using both dies of SRAM memory, this ability has now been added. Will need some additional testing to make sure all works.