FLASH bugfix; error counting moved; memories expanded
There was a bug with the FLASH that it could not write to more than one page of the memory; this has been troubleshooted and resolved. Error counting has been moved in to the error creation and deletion functions for simplification. Memories which previously were limited for debugging have been allowed to use their maximum EXCEPT FLASH because it takes forever.